refactor: swift 파일 폴더 재배치 enum, protocol, error 폴더를 만들어서 파일 재배치
