+++ OpenMarket/OpenMarket/Manager/ParsingManager.swift
@@ -7,23 +7,6 @@
 
 import UIKit
 
-enum ParsingError: Error, LocalizedError {
-    case assetLoadFailed
-    case decodingFailed
-    case encodingFailed
-    
-    var errorDescription: String {
-        switch self {
-        case .assetLoadFailed:
-            return "에셋 데이터를 불러오는데 실패했습니다."
-        case .decodingFailed:
-            return "디코딩에 실패했습니다."
-        case .encodingFailed:
-            return "인코딩에 실패했습니다."
-        }
-    }
-}
-
 struct ParsingManager {
     private let jsonDecoder = JSONDecoder()
